+++ src/commands/org/open/agent.ts
@@ -53,11 +53,11 @@ export class OrgOpenAgent extends OrgOpenCommandBase<OrgOpenOutput> {
     this.connection = this.org.getConnection(flags['api-version']);
 
     const [frontDoorUrl, retUrl] = await Promise.all([
-      buildFrontdoorUrl(this.org, this.connection),
+      buildFrontdoorUrl(this.org, this.connection, true),
       buildRetUrl(this.connection, flags.name),
     ]);
 
-    return this.openOrgUI(flags, frontDoorUrl, retUrl);
+    return this.openOrgUI(flags, frontDoorUrl, encodeURIComponent(retUrl));
   }
 }