+++ CHANGELOG.md
@@ -1,3 +1,83 @@
+## [5.5.8](https://github.com/salesforcecli/plugin-org/compare/5.5.7...5.5.8) (2025-04-06)
+
+### Bug Fixes
+
+- **deps:** bump @salesforce/source-deploy-retrieve ([bb86872](https://github.com/salesforcecli/plugin-org/commit/bb868724306d3930e6be121e002fd3b1ceb800ba))
+
+## [5.5.7](https://github.com/salesforcecli/plugin-org/compare/5.5.6...5.5.7) (2025-03-30)
+
+### Bug Fixes
+
+- **deps:** bump @salesforce/core from 8.8.6 to 8.8.7 ([d08a2c8](https://github.com/salesforcecli/plugin-org/commit/d08a2c8e45b8a4582950667a0bfd886ec8eb7549))
+
+## [5.5.6](https://github.com/salesforcecli/plugin-org/compare/5.5.5...5.5.6) (2025-03-27)
+
+### Bug Fixes
+
+- correct retURL to startURL for OTP URLs W-18077630 ([#1387](https://github.com/salesforcecli/plugin-org/issues/1387)) ([57376f4](https://github.com/salesforcecli/plugin-org/commit/57376f40c28654fd0290a31117b2f61ab29bdbbf))
+
+## [5.5.5](https://github.com/salesforcecli/plugin-org/compare/5.5.4...5.5.5) (2025-03-24)
+
+### Bug Fixes
+
+- **deps:** bump @salesforce/core from 8.8.5 to 8.8.6 ([42e5b28](https://github.com/salesforcecli/plugin-org/commit/42e5b2896c02fd4140569e59793b5d62aaafc257))
+
+## [5.5.4](https://github.com/salesforcecli/plugin-org/compare/5.5.3...5.5.4) (2025-03-23)
+
+### Bug Fixes
+
+- **deps:** bump @salesforce/source-deploy-retrieve ([18fa0cb](https://github.com/salesforcecli/plugin-org/commit/18fa0cb359bb0fc4bfa6e1651e836b45a371c966))
+
+## [5.5.3](https://github.com/salesforcecli/plugin-org/compare/5.5.2...5.5.3) (2025-03-23)
+
+### Bug Fixes
+
+- **deps:** bump @salesforce/sf-plugins-core from 12.2.0 to 12.2.1 ([e18e7e0](https://github.com/salesforcecli/plugin-org/commit/e18e7e0acada146454b6829927cc7023fc80dd8f))
+
+## [5.5.2](https://github.com/salesforcecli/plugin-org/compare/5.5.1...5.5.2) (2025-03-23)
+
+### Bug Fixes
+
+- **deps:** bump @oclif/multi-stage-output from 0.8.9 to 0.8.12 ([a30131c](https://github.com/salesforcecli/plugin-org/commit/a30131c4d6a4784ebd867231145bacc47704ca78))
+
+## [5.5.1](https://github.com/salesforcecli/plugin-org/compare/5.5.0...5.5.1) (2025-03-19)
+
+### Bug Fixes
+
+- update warning msg ([#1381](https://github.com/salesforcecli/plugin-org/issues/1381)) ([0766ad8](https://github.com/salesforcecli/plugin-org/commit/0766ad859e30ed0b36cc4bf1b8218dbc1e4893a3))
+
+# [5.5.0](https://github.com/salesforcecli/plugin-org/compare/5.4.1...5.5.0) (2025-03-18)
+
+### Features
+
+- use UI Bridge API to generate a single-use frontdoor ([#1375](https://github.com/salesforcecli/plugin-org/issues/1375)) ([a9bf347](https://github.com/salesforcecli/plugin-org/commit/a9bf347706e3aecfc72385b1b5730471e63f5b13))
+
+## [5.4.1](https://github.com/salesforcecli/plugin-org/compare/5.4.0...5.4.1) (2025-03-16)
+
+### Bug Fixes
+
+- **deps:** bump @oclif/core from 4.2.8 to 4.2.10 ([7e15eb0](https://github.com/salesforcecli/plugin-org/commit/7e15eb04258d6a591f254056761559477de3cf55))
+
+# [5.4.0](https://github.com/salesforcecli/plugin-org/compare/5.3.0...5.4.0) (2025-03-12)
+
+### Features
+
+- use MSO in sandbox commands ([#1297](https://github.com/salesforcecli/plugin-org/issues/1297)) ([c1e706b](https://github.com/salesforcecli/plugin-org/commit/c1e706b6060b76d5a6c273fd38c022d90449d1fd))
+
+# [5.3.0](https://github.com/salesforcecli/plugin-org/compare/5.2.42...5.3.0) (2025-03-11)
+
+### Bug Fixes
+
+- add the exclusive flag to edition and source-org ([7f58889](https://github.com/salesforcecli/plugin-org/commit/7f58889e11f09fe4d35e5a8a5919d18dc62cdd9e))
+- more help updates ([cd921c6](https://github.com/salesforcecli/plugin-org/commit/cd921c6b9af16b3a4b2d312ceb9e6e9ebdb29b66))
+- update help with info about new --snapshot flag ([b01a47a](https://github.com/salesforcecli/plugin-org/commit/b01a47a242922b8ed22651e8dd7c7d7f246d10ea))
+- update the changes ([4678a9a](https://github.com/salesforcecli/plugin-org/commit/4678a9abfa496c5fab254c6e3663c2806010ffe5))
+
+### Features
+
+- add a --snapshot flag for creating scratch org ([4566d6d](https://github.com/salesforcecli/plugin-org/commit/4566d6d9ddfa7fc0fec81b801e424c207ff187fd))
+- add test for snapshot ([e8baad1](https://github.com/salesforcecli/plugin-org/commit/e8baad127e92be1a0b59fba6c840617bb48d4829))
+
 ## [5.2.42](https://github.com/salesforcecli/plugin-org/compare/5.2.41...5.2.42) (2025-03-09)
 
 ### Bug Fixes